Project Completion

+++ src/main/java/DTO/TransactionDTO.java
@@ -1,11 +1,11 @@
 package DTO;
 
 public class TransactionDTO implements DTO {
-    public final int itemCode;
-    public final int paymentMethod;
+    public final long itemCode;
+    public final long paymentMethod;
     public final long machineCode;
 
-    public TransactionDTO(int machineCode, int itemCode, int paymentMethod) {
+    public TransactionDTO(long machineCode, long itemCode, long paymentMethod) {
         this.machineCode = machineCode;
         this.itemCode = itemCode;
         this.paymentMethod = paymentMethod;
